cdk/ec2_instance/requirements.txt
blinker==1.6.2
boto3==1.28.23
botocore==1.31.23
click==8.1.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
